requirements.txt
requests
tqdm
python-dotenv
tenacity
colorama
textual>=0.40.0
